implement the add meal functionality for the admin.

+++ app/src/main/java/com/mona/adel/quickbite/data/relations/MealDayCrossRef.kt
@@ -8,6 +8,6 @@ import androidx.room.Entity
     primaryKeys = ["mealId", "dayId"]
 )
 data class MealDayCrossRef(
-    var mealId: Int,
+    var mealId: Long,
     var dayId: Int
 )
